Category 페이지의 CategoryBoard와 KeywordBoard를 파일로 분리
